Add timeout to prepare for mount/dismount
After the robot has been told to prepare for mount/dismount we require a
timeout routine to run if no further instructions are received. This is
to prevent the robot remaining in the LN2 dewar and icing up.

We achieve this by spawning a new thread that waits for a timeout abort
event. The mount/dismount functions will set this event if they are
called. If the event is not set within a timeout the prepare_timeout
function will start a task to send the robot home.